CRAN submission changes (#189)
* Add cran-comments.md

* Fix CRAN check errors

* Additional checks for tests

* Address CRAN submission feedback

* Regenerate site documentation

* Bump upload-artifact

* Add CRAN badge & regenerate site documentation

+++ tests/testthat/test-ResultsDataModel.R
@@ -34,15 +34,16 @@ sqliteResultsDatabaseSchema <- "main"
 
 withr::defer(
   {
-    connection <- DatabaseConnector::connect(connectionDetails = postgresConnectionDetails)
-    sql <- "DROP SCHEMA IF EXISTS @resultsDatabaseSchema CASCADE;"
-    DatabaseConnector::renderTranslateExecuteSql(
-      sql = sql,
-      resultsDatabaseSchema = postgresResultsDatabaseSchema,
-      connection = connection
-    )
-
-    DatabaseConnector::disconnect(connection)
+    if (Sys.getenv("CDM5_POSTGRESQL_SERVER") != "") {
+      connection <- DatabaseConnector::connect(connectionDetails = postgresConnectionDetails)
+      sql <- "DROP SCHEMA IF EXISTS @resultsDatabaseSchema CASCADE;"
+      DatabaseConnector::renderTranslateExecuteSql(
+        sql = sql,
+        resultsDatabaseSchema = postgresResultsDatabaseSchema,
+        connection = connection
+      )
+      DatabaseConnector::disconnect(connection)
+    }
     unlink(databaseFile, force = TRUE)
   },
   testthat::teardown_env()
@@ -81,6 +82,7 @@ testCreateSchema <- function(connectionDetails, resultsDatabaseSchema) {
 }
 
 test_that("Create schema", {
+  skip_on_cran()
   testCreateSchema(
     connectionDetails = postgresConnectionDetails,
     resultsDatabaseSchema = postgresResultsDatabaseSchema
@@ -126,6 +128,7 @@ testUploadResults <- function(connectionDetails, resultsDatabaseSchema, resultsF
 }
 
 test_that("Results upload", {
+  skip_on_cran()
   unzipFolder <- tempfile("unzipTempFolder", tmpdir = tempdir())
   dir.create(path = unzipFolder, recursive = TRUE)
   on.exit(unlink(unzipFolder, recursive = TRUE), add = TRUE)
